feat(connector): force done to be called with an object
This breaking change will solve an issue with getting the correct
value out of the connector when coming to validation. Multiple variables
can be used to drive a component, but only one of them may be needed as
the validated variable. For example, the token input requires a state
for the input, and another for the chips - however only the chips are
important at the end of the day.