feat(hoc): make setting up forms simpler
This commit makes dealing with converting between state values and the
values you wish to get out of a form a problem for each component's
connector. Using a .toValue and a .getDefault method on the connector,
we can easily go between a value fit for the store and a value fit for
consumption.

As a benefit, this also means we can set up forms dynamically using each
field's connector. It's now possible to set up a form with just an empty
fields and error object and it will be created on component load.