refactor(*): rewrite to remove antipatterns
This refactor is a huge change, but the fundamentals are the same.
Instead of having all the functionality within the react component, the
logic has all been moved into actions. This means that the developer now
has full control over the flow of the validation.

The clean method has been removed, as actions can directly change the
output now. Doing this means the clean function has full access to the
cerebral ecosystem. An even bigger bonus is that the clean function can
be made asyncronous with no extra work involved.

For this system to work, forms now have to be named and registered
globally so that they can be accessed in the cerebral services.